feat: 댓글 목록 조회 API 코드 일부를 모듈화 추가로 commentService.js에서 쓰던 fillterOptions 코드를 다른 파일로 이동
