+++ express-mission-prisma/src/utils/buildCursorResponse.js
@@ -1,0 +1,17 @@
+function buildCursorResponse(list, total, pageSize) {
+  const lastList = list[pageSize];
+  const NextCusor = lastList ? lastList.id : "null";
+  if (NextCusor !== "null") {
+    list.pop();
+  }
+
+  return {
+    cursorInfo: {
+      total,
+      NextCusor,
+    },
+    list,
+  };
+}
+
+export default buildCursorResponse;
